refactor: move tests for reuse
