+++ test/api/lxc/create_lxc_test.go
@@ -1,4 +1,4 @@
-package api
+package lxc
 
 import (
 	"context"
