+++ test/api/lxc/helper.go
@@ -1,4 +1,4 @@
-package api
+package lxc
 
 import (
 	"context"
