+++ test/api/qemu/clone_qemu_test.go
@@ -1,4 +1,4 @@
-package api
+package qemu
 
 import (
 	"context"
