+++ test/api/qemu/config.go
@@ -1,4 +1,4 @@
-package api
+package qemu
 
 import (
 	"strings"
