+++ test/api/qemu/helper.go
@@ -1,4 +1,4 @@
-package api
+package qemu
 
 import (
 	"context"
